# Optimize counter inc-by-one