--- a/spectator/meter/counter.go
+++ b/spectator/meter/counter.go
@@ -13,14 +13,14 @@ import (
 //
 // https://netflix.github.io/spectator/en/latest/intro/counter/
 type Counter struct {
-	id              *Id
-	writer          writer.Writer
-	meterTypeSymbol string
+	id           *Id
+	writer       writer.Writer
+	incByOneLine string
 }
 
 // NewCounter generates a new counter, using the provided meter identifier.
 func NewCounter(id *Id, writer writer.Writer) *Counter {
-	return &Counter{id, writer, "c"}
+	return &Counter{id, writer, fmt.Sprintf("c:%s:1", id.spectatordId)}
 }
 
 // MeterId returns the meter identifier.
@@ -30,14 +30,17 @@ func (c *Counter) MeterId() *Id {
 
 // Increment increments the counter.
 func (c *Counter) Increment() {
-	var line = fmt.Sprintf("%s:%s:%d", c.meterTypeSymbol, c.id.spectatordId, 1)
-	c.writer.Write(line)
+	c.writer.Write(c.incByOneLine)
 }
 
 // Add adds an int64 delta to the current measurement.
 func (c *Counter) Add(delta int64) {
+	if delta == 1 {
+		c.writer.Write(c.incByOneLine)
+		return
+	}
 	if delta > 0 {
-		var line = fmt.Sprintf("%s:%s:%d", c.meterTypeSymbol, c.id.spectatordId, delta)
+		var line = fmt.Sprintf("c:%s:%d", c.id.spectatordId, delta)
 		c.writer.Write(line)
 	}
 }
@@ -45,7 +48,7 @@ func (c *Counter) Add(delta int64) {
 // AddFloat adds a float64 delta to the current measurement.
 func (c *Counter) AddFloat(delta float64) {
 	if delta > 0.0 {
-		var line = fmt.Sprintf("%s:%s:%f", c.meterTypeSymbol, c.id.spectatordId, delta)
+		var line = fmt.Sprintf("c:%s:%f", c.id.spectatordId, delta)
 		c.writer.Write(line)
 	}
 }