Conceivably increment-by-one is by far the most common operation on counters. This PR optimize such operation, making it alloc-free.

Benchmark with:
```
func BenchmarkInc(b *testing.B) {
	b.Run("Increment", func(b *testing.B) {
		w := writer.MemoryWriter{}
		id := NewId("foo", nil)
		c := NewCounter(id, &w)
		b.ReportAllocs()
		b.ResetTimer()
		for i := 0; i < b.N; i++ {
			c.Increment()
		}
	})
}
```

before optimization:
```
goos: darwin
goarch: arm64
pkg: github.com/Netflix/spectator-go/v2/spectator/meter
cpu: Apple M3 Pro
BenchmarkInc/Increment-12               10298648               115.2ns/op           120 B/op          3 allocs/op
PASS
ok      github.com/Netflix/spectator-go/v2/spectator/meter      2.485s
```

after optimization:
```
goos: darwin
goarch: arm64
pkg: github.com/Netflix/spectator-go/v2/spectator/meter
cpu: Apple M3 Pro
BenchmarkInc/Increment-12               40374824                31.44ns/op           97 B/op          0 allocs/op
PASS
ok      github.com/Netflix/spectator-go/v2/spectator/meter      3.563s
```