@dmuino 2024-11-12
I like the idea in principle but ideally we'd also like to understand the effect of the additional branching where a mixture of increment by one, and increment by other numbers might introduce some stalling.

Is this change motivated by a concrete problem or is it more abstract in nature? In particular making counters significantly different than the other meter types is adding some complexity to the code that needs to be weighed properly.